requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.12.0
pydantic-settings>=2.12.0

# AI/ML Libraries
openai-whisper==20231117
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers==4.35.2
accelerate==0.25.0
sentencepiece>=0.1.99  # Installed: 0.2.1 (use newer version to avoid build issues)

# TTS
gTTS==2.4.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Optional (for better performance)
# onnxruntime-gpu==1.16.3  # Uncomment if you have GPU

